mwaa-plugin/requirements.txt
# Core dependencies
apache-airflow>=2.0.0
croniter>=0.3
flask_admin>=1.5.8
flask_appbuilder
flask_babel

# Provider-specific dependencies
apache-airflow-providers-microsoft-azure>=1.1.0
azure-storage-blob
apache-airflow-providers-databricks
jaydebeapi
apache-airflow-providers-jdbc
apache-airflow-providers-snowflake>=2.0.0
apache-airflow-providers-google